refactored login folder
